refactor(themes): Rearranged files
