+++ src/themes/BlueBaseTheme.ts
@@ -1,3 +1,3 @@
-import { Theme } from '../structure';
+import { Theme } from './Theme';
 
 export const BlueBaseTheme = new Theme();
