--- src/themes/helpers/index.ts
@@ -1,4 +1,0 @@
-export * from './applyStyles';
-export * from './createPalette';
-export * from './createThemeVariant';
-export * from './createTypography';
