+++ src/themes/index.ts
@@ -1,4 +1,6 @@
-export * from './helpers';
-export * from './structure';
-export * from './themes';
+export * from './applyStyles';
+export * from './BlueBaseTheme';
+export * from './Palette';
+export * from './Theme';
 export * from './ThemeContext';
+export * from './Typography';
