--- src/themes/structure/index.ts
@@ -1,3 +1,0 @@
-export * from './Palette';
-export * from './Theme';
-export * from './Typography';
