--- src/themes/themes/index.ts
@@ -1,1 +1,0 @@
-export * from './BlueBaseTheme';
